Fix token_to_expert: use register_buffer instead of Parameter for torch.compile compat
nn.Parameter with float tensor caused pickle serialization errors in
torch.compile's AOT autograd cache. Use register_buffer (persistent=True)
with long dtype instead.

+++ vllm/model_executor/layers/token_routed_i64/fused_i64_moe.py
@@ -97,7 +97,7 @@ def i64_token_routed_forward(
     """
     # === I64 Routing (inside splitting op = eager) ===
     token_ids_clamped = token_ids.clamp(0, vocab_size - 1)
-    expert_ids = token_to_expert[token_ids_clamped].long()
+    expert_ids = token_to_expert[token_ids_clamped]
 
     # Mu-guided bias — always compute (no CPU sync like .any())
     # When mu is zeros, mu_logits is zeros, base_one_hot * 10.0 dominates,
